requirements.txt
aliyun-python-sdk-core-v3==2.9.0
aliyun-python-sdk-dysmsapi==1.0.0
atomicwrites==1.1.5
attrs==18.2.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
confusable-homoglyphs==2.0.2
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.15
django-allauth==0.34.0
django-appconf==1.0.2
django-braces==1.13.0
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-grappelli==2.11.1
django-guardian==1.4.9
django-jet==1.0.7
django-messages==0.5.3
django-model-utils==3.0.0
django-modeladmin-reorder==0.3.1
django-oauth-toolkit==1.0.0
django-private-chat==0.1.9
django-redis==4.9.0
django-registration==2.3
django-rest-auth==0.9.3
django-rest-swagger==2.0.7
django-users2==0.2.1
djangorestframework==3.7.7
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
drf-nested-routers==0.90.0
et-xmlfile==1.0.1
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
more-itertools==4.3.0
mysqlclient==1.3.12
oauthlib==2.0.6
olefile==0.44
openapi-codec==1.3.2
openpyxl==2.5.6
Pillow==4.0.0
pinax-messages==2.0.2
pluggy==0.7.1
py==1.6.0
pycryptodome==3.6.4
pycryptodomex==3.6.6
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.8.0
pytest==3.7.1
python-alipay-sdk==1.8.0
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.9.0
South==1.0.2
submail==0.4
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
websockets==3.2
